feat: migrate from homebridge node-pty to Microsoft node-pty v1.1.0-beta34
- Replace @homebridge/node-pty-prebuilt-multiarch with official node-pty from Microsoft
- Build node-pty from source using latest beta tag (v1.1.0-beta34)
- Create SEA-compatible patches managed by pnpm patch system
- Simplify build-native.js by removing inline patching logic
- Update all imports and build scripts
- Add VIBETUNNEL_SEA environment variable for SEA builds

This migration ensures we're using the latest upstream node-pty with full
control over the build process and patches.
